moving merkle pieces under merkle mod
