+++ zkvm/src/utreexo/mod.rs
@@ -9,8 +9,8 @@ mod tests;
 
 // Public API
 pub use self::forest::{Catchup, Forest, UtreexoError, WorkForest};
-pub use self::path::{Path, Position, Proof};
-pub use super::merkle::Hasher;
+pub use self::path::{Proof};
+pub use super::merkle::{Hasher};
 
 /// Utreexo-labeled hasher.
 pub fn utreexo_hasher<T: super::merkle::MerkleItem>() -> Hasher<T> {
